Switch to BNLS_VERSIONCHECKEX2 (0x1A)
Switching to BNLS_VERSIONCHECKEX2 to provide BNLS with the full
name of the checkrevision archive. Previously, an issue was
preventing the program from connecting to a PvPGN server if
going through JBLS using BNLS_VERSIONCHECK (0x09) as JBLS
was assuming official Battle.Net data while PvPGN was
providing different data.